+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2026-01-18
+
 ### Added
 
 - `wtm config edit` command to open project configuration in `$EDITOR`
@@ -68,7 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shell integration examples (wtm-cd, fzf integration)
 - MIT License
 
-[Unreleased]: https://github.com/choplin/wtm/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/choplin/wtm/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/choplin/wtm/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/choplin/wtm/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/choplin/wtm/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/choplin/wtm/compare/v0.1.1...v0.3.0